Roboshop

+++ Roboshop/components/frontend.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
+source ./common.sh
 
 set -e
-echo "hello"
-yum install nginx -y
-systemctl enable nginx
-systemctl start nginx
+echo "hello I'm front end"
+
+userid = "$(id -u)"
+if[$userid -ne 0];then
+    echo "You need to be a root user"
+    exit 1
+fi
+
 curl -s -L -o /tmp/frontend.zip "https://github.com/stans-robot-project/frontend/archive/main.zip"
 cd /usr/share/nginx/html
 rm -rf *
